Finally, a fully working multi-gunner system
- Added a fully working multi-gunner seat system to the Xylvanian Bomber as a test

The X Bomber is the only model that has the necessary assets for the multi-gunner system atm. I have to redo several parts on all of the models anyway, so now I have to add the proper assets to the other models to support the new gunner system. Hopefully you are excited for more BW content soon!